Add NER model variant with required fields
In order to have an NER model that is simpler for internal regex/CFG
representations, add an NER variant that requires all fields and does
not include a default value.

In particular, this makes it possible to evaluate a version of NER for
Outlines and provides an additional point of comparison for other
libraries.